refactor: 예외 처리 클래스 분리
