--- src/test/java/cart/JwpCartApplicationTests.java
@@ -1,13 +1,0 @@
-package cart;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JwpCartApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
